requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
python-socketio==5.10.0
requests==2.31.0
